requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
mysqlclient>=2.0.0
wheel==0.45.1
